Refactoring to classes
- Create one class by Model
- Use MLFlow by default in each class
- Expose retraining methods
- Update FastApi
