+++ detect_ai_content/ml_logic/evaluation.py
@@ -1,0 +1,12 @@
+
+from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score
+
+def evaluate_model(model, X_test_processed, y_test):
+    y_pred = model.predict(X_test_processed)
+
+    return {
+        "accuracy_score": accuracy_score(y_true=y_test, y_pred=y_pred),
+        "f1_score" : f1_score(y_true=y_test, y_pred=y_pred),
+        "precision_score": precision_score(y_true=y_test, y_pred=y_pred),
+        "recall_score" : recall_score(y_true=y_test, y_pred=y_pred)
+    }
