+++ detect_ai_content/ml_logic/for_texts/using_ml_features/TrueNetTextLogisticRegression.py
@@ -1,7 +1,12 @@
 
+import pickle
+import os
+import mlflow
+from mlflow import MlflowClient
+
+from sklearn.linear_model import LogisticRegression
 from sklearn.preprocessing import RobustScaler
-from sklearn.svm import SVC
-from sklearn.model_selection import train_test_split, GridSearchCV
+from sklearn.model_selection import train_test_split
 from sklearn.pipeline import make_pipeline
 
 from detect_ai_content.ml_logic.mlflow import mlflow_save_metrics, mlflow_save_model, mlflow_save_params, load_model
@@ -9,80 +14,55 @@ from detect_ai_content.params import *
 from detect_ai_content.ml_logic.data import get_enriched_df
 from detect_ai_content.ml_logic.evaluation import evaluate_model
 
-import os
-import pickle
-import mlflow
-from mlflow import MlflowClient
-
-class TrueNetTextSVC:
+class TrueNetTextLogisticRegression:
     def _load_model(self):
         """
         Model sumary :
-            Trained TBD
-            Algo : TBD
-            Cross Validate average result (0.2 test) : TBD
+            Trained in 2,532,099 texts (using 3 datasets combined)
+            Algo : LogisticRegression
+            Cross Validate average result (0.2 test) : 0.83
         """
         return load_model(self.mlflow_model_name, is_tensorflow=False, stage="Production")
 
     def __init__(self):
-        self.name = "TrueNetTextSVC"
+        self.name = "TrueNetTextLogisticRegression"
         self.description = ""
-        self.mlflow_model_name = "TrueNetTextSVC"
-        self.mlflow_experiment = "TrueNetTextSVC_experiment_leverdewagon"
+        self.mlflow_model_name = "TrueNetTextLogisticRegression"
+        self.mlflow_experiment = "TrueNetTextLogisticRegression_experiment_leverdewagon"
         self.model = self._load_model()
 
-    def run_grid_search():
-        param_grid = {
-            'C':[1,10,100,1000],
-            'gamma':[1,0.1,0.001,0.0001],
-            'kernel':['linear','rbf']
-        }
-
-        df = get_enriched_df(10_000)
-        X = df[[
-            'repetitions_ratio',
-            'punctuations_ratio',
-            'text_corrections_ratio',
-            'average_sentence_lenght',
-            'average_neg_sentiment_polarity',
-        ]]
-        y = df['generated']
-        grid = GridSearchCV(SVC(),param_grid,refit = True, verbose=2)
-        grid.fit(X,y)
-        print(grid.best_estimator_)
-
     def retrain_full_model():
+        print("retrain_full_model START")
+        import detect_ai_content
+        module_dir_path = os.path.dirname(detect_ai_content.__file__)
 
         # init
         client = MlflowClient()
-        futur_obj = TrueNetTextSVC()
+        futur_obj = TrueNetTextLogisticRegression()
         experiment_id = client.get_experiment_by_name(futur_obj.mlflow_experiment).experiment_id
         mlflow.start_run(experiment_id=experiment_id)
 
+        big_df = get_enriched_df()
 
-        big_df = get_enriched_df(10_000)
-
-        pipeline = make_pipeline(
+        pipeline_linear_regression = make_pipeline(
             RobustScaler(),
-            SVC(C=10, gamma=0.1) # param from run_grid_search
+            LogisticRegression()
         )
 
         X = big_df[[
-            'repetitions_ratio',
-            'punctuations_ratio',
-            'text_corrections_ratio',
-            'average_sentence_lenght',
-            'average_neg_sentiment_polarity',
+        'repetitions_ratio',
+        'punctuations_ratio',
+        'text_corrections_ratio',
+        'average_sentence_lenght',
+        'average_neg_sentiment_polarity',
         ]]
         y = big_df['generated']
 
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
 
-        model = pipeline.fit(X_train, y_train)
+        model = pipeline_linear_regression.fit(X=X_train, y=y_train)
 
-        import detect_ai_content
-        module_dir_path = os.path.dirname(detect_ai_content.__file__)
-        model_path = f = f'{module_dir_path}/../detect_ai_content/models/leverdewagon/{futur_obj.mlflow_model_name}.pickle'
+        model_path = f'{module_dir_path}/../detect_ai_content/models/leverdewagon/{futur_obj.mlflow_model_name}.pickle'
         pickle.dump(model, open(model_path, 'wb'))
 
         # mlflow_save_params
