+++ detect_ai_content/ml_logic/mlflow.py
@@ -1,13 +1,22 @@
 
-import mlflow
+
 from detect_ai_content.params import *
+
+import mlflow
+mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
 
 def load_model(model_name: str, is_tensorflow: bool, stage:str):
     model_uri = f"models:/{model_name}/{stage}"
-    if is_tensorflow:
-        model = mlflow.tensorflow.load_model(model_uri=model_uri)
-    else:
-        model = mlflow.sklearn.load_model(model_uri=model_uri)
+
+    try:
+        if is_tensorflow:
+            model = mlflow.tensorflow.load_model(model_uri=model_uri)
+        else:
+            model = mlflow.sklearn.load_model(model_uri=model_uri)
+
+    except mlflow.exceptions.MlflowException:
+        print(f"Oops!  there is no model:{model_name} on stage:{stage}")
+        model = None
     return model
 
 def mlflow_save_params(
