+++ detect_ai_content/ml_logic/preprocess.py
@@ -1,23 +1,27 @@
 import pandas as pd
 
-def text_lenght(text):
-    return len(text)
+from detect_ai_content.ml_logic.data import enrich_text
+from sklearn.preprocessing import RobustScaler
 
-def average_sentences(text):
-    return len(text.split("."))
+def preprocess(data, auto_enrich=True):
+    if auto_enrich:
+        data_processed = enrich_text(data=data.copy())
+    else:
+        data_processed = data.copy()
 
-def preprocess_text(text):
-    X = pd.DataFrame(dict(
-        text=[text],
-    ))
+    data_processed['repetitions_ratio'] = data_processed['text_repetitions_nb']/data_processed['text_lenght']
+    data_processed['punctuations_ratio'] = data_processed['punctuations_nb']/data_processed['text_lenght']
+    data_processed['text_corrections_ratio'] = data_processed['text_corrections_nb']/data_processed['text_lenght']
+    data_processed['average_sentence_lenght'] = data_processed['text_lenght']/data_processed['number_of_sentences']
+    data_processed['average_neg_sentiment_polarity'] = data_processed['neg_sentiment_polarity']/data_processed['text_lenght']
 
-    X_processed = preprocess_features(X)
-    print(X_processed)
-    return X_processed
+    data_processed = data_processed[[
+        'repetitions_ratio',
+        'punctuations_ratio',
+        'text_corrections_ratio',
+        'average_sentence_lenght',
+        'average_neg_sentiment_polarity'
+    ]]
 
-def preprocess_features(df):
-    texts_df = df.copy()
-    texts_df['text_lenght'] = texts_df['text'].apply(text_lenght)
-    texts_df['sentences_count'] = texts_df['text'].apply(average_sentences)
-    texts_df = texts_df.drop(columns=['text'])
-    return texts_df
+    scaler = RobustScaler()
+    return scaler.fit_transform(data_processed)
