+++ detect_ai_content/params.py
@@ -1,12 +1,6 @@
-import os
-import numpy as np
 
 MODEL_TARGET="mlflow"
 MLFLOW_TRACKING_URI="https://mlflow.lewagon.ai"
-
-# Manual feature engineering model
-MLFLOW_FE_EXPERIMENT="TrueNet_FE_experiment_leverdewagon"
-MLFLOW_FE_MODEL_NAME="TrueNet_FE_detect_ai_content"
 
 # Vectorizer model
 MLFLOW_VECTORIZER_MODEL_NAME="TrueNet_Tfid_NaiveBayes_detect_ai_content"
