+++ tests/test_texts_mlflow_predict.py
@@ -1,27 +1,24 @@
 
 import unittest
-import mlflow
 import pandas as pd
 
 from detect_ai_content.ml_logic.mlflow import load_model
 from detect_ai_content.params import *
-from detect_ai_content.ml_logic.for_texts.using_ml_features.using_ml_features import preprocess, evaluate_model
+from detect_ai_content.ml_logic.for_texts.using_ml_features.TrueNetTextLogisticRegression import *
+from detect_ai_content.ml_logic.for_texts.using_ml_features.TrueNetTextTfidfNaiveBayesClassifier import *
+from detect_ai_content.ml_logic.preprocess import preprocess
 
 class TestMLFlowTextPrediction(unittest.TestCase):
     def test_load_text_model_FE(self):
-        mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
-        model = load_model(model_name=MLFLOW_FE_MODEL_NAME, is_tensorflow=False, stage="Production")
+        model = TrueNetTextLogisticRegression().model
         self.assertEqual(model is not None, True)
 
     def test_load_text_model_Vectorizer(self):
-        mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
-        model = load_model(model_name=MLFLOW_VECTORIZER_MODEL_NAME, is_tensorflow=False, stage="Production")
+        model = TrueNetTextTfidfNaiveBayesClassifier().model
         self.assertEqual(model is not None, True)
 
     def test_mlflow_model_FE_predictions(self):
-        mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
-        model = load_model(model_name=MLFLOW_FE_MODEL_NAME, is_tensorflow=False, stage="Production")
-
+        model = TrueNetTextLogisticRegression().model
         path = "./tests/data/sample_dataset_1000.csv"
         df = pd.read_csv(path)
         X = df[['text']]
