+++ tests/test_texts_modeling_using_ml_features.py
@@ -1,10 +1,9 @@
 
 import unittest
 import pandas as pd
-from sklearn.model_selection import train_test_split
 
-from detect_ai_content.ml_logic.for_texts.using_ml_features.using_ml_features import preprocess,enrich, train_LogisticRegression_model, evaluate_model
 from detect_ai_content.ml_logic.for_texts.using_ml_features.features.features import compute_number_of_text_corrections, compute_number_of_text_corrections_using_TextBlob, compute_number_of_text_corrections_using_pyspellchecker, compute_number_of_text_corrections_using_nltk_words
+from detect_ai_content.ml_logic.preprocess import preprocess
 
 def get_human_texts():
     texts = [
@@ -70,28 +69,6 @@ Students fall in environment learning because they learn at class when the teach
         X = df[['text']]
         X_preprocessed = preprocess(X)
         self.assertEqual("text" in X_preprocessed, False)
-
-    def test_training(self):
-        texts = get_human_texts()
-        human_df = pd.DataFrame(data={"text": texts})
-        human_df['generated'] = 0
-
-        texts = get_ai_texts()
-        ai_df = pd.DataFrame(data={"text": texts})
-        ai_df['generated'] = 1
-
-        df = pd.concat(objs=[ai_df, human_df])
-        X = df[['text']]
-        y = df['generated']
-
-        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
-
-        X_train_preprocessed = preprocess(X_train)
-        model = train_LogisticRegression_model(X_train_preprocessed, y_train)
-
-        X_test_preprocessed = preprocess(X_test)
-        results = evaluate_model(model=model, X_test_processed=X_test_preprocessed, y_test=y_test)
-        print(results)
 
 if __name__ == '__main__':
     unittest.main()
